# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Town Backend",
    "description": "二次元 AI 小镇陪伴智能体 - World Engine + LLM",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tools/servers/{server_name}": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Get Tool Server Detail",
        "description": "获取单个命名空间详情\n\nArgs:\n    server_name: 命名空间名称（如 \"shop\"）",
        "operationId": "get_tool_server_detail_api_v1_tools_servers__server_name__get",
        "parameters": [
          {
            "name": "server_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tools/servers/{server_name}/enabled": {
      "put": {
        "tags": [
          "tools"
        ],
        "summary": "Toggle Tool Server",
        "description": "启用/禁用整个命名空间的所有工具\n\n状态持久化到 Redis hash `tools:enabled`，Character Tick 决策时\n会读取该状态过滤可用工具列表。\n\nArgs:\n    server_name: 命名空间名称\n    payload: {\"enabled\": true|false}\n\nReturns:\n    更新后的启用状态",
        "operationId": "toggle_tool_server_api_v1_tools_servers__server_name__enabled_put",
        "parameters": [
          {
            "name": "server_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerToggleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/tools/{tool_name}/invoke": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Invoke Tool",
        "description": "调用本地工具（测试用）\n\nArgs:\n    tool_name: 工具全名（如 \"shop.buy_item\"）或简短名（如 \"buy_item\"）\n    server_name: 可选命名空间（用于消歧）\n    args: 工具参数（JSON body）\n\nReturns:\n    工具执行结果",
        "operationId": "invoke_tool_api_v1_tools_tools__tool_name__invoke_post",
        "parameters": [
          {
            "name": "tool_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Name"
            }
          },
          {
            "name": "server_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Server Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Args"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolInvokeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ServerDetailOut": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "ServerDetailOut",
        "description": "工具命名空间详情（tools 列表动态，透传保留）"
      },
      "ServerToggleOut": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "server": {
            "type": "string",
            "title": "Server"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "success",
          "server",
          "enabled"
        ],
        "title": "ServerToggleOut"
      },
...
      "ToolInvokeOut": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "success",
          "endpoint"
        ],
        "title": "ToolInvokeOut",
        "description": "工具调用结果（result 结构随工具各异，透传保留）"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}